+++ src/ascent/runtimes/ascent_main_runtime.hpp
@@ -80,7 +80,6 @@ public:
 
     void  Publish(const conduit::Node &data) override;
     void  Execute(const conduit::Node &actions) override;
-
     void  Info(conduit::Node &out) override;
 
     void  Cleanup() override;
@@ -116,6 +115,7 @@ private:
 
     std::string       m_session_name;
     conduit::Node     m_save_session_actions;
+    conduit::Node     m_save_info_actions;
 
     bool              m_field_filtering;
     std::set<std::string> m_field_list;
@@ -123,6 +123,7 @@ private:
     conduit::Node     m_comments;
 
     void              ResetInfo();
+    void              AddPublishedMeshInfo();
 
     flow::Workspace w;
     conduit::Node CreateDefaultFilters();
@@ -150,7 +151,13 @@ private:
     void SourceFieldFilter();
     void PaintNestsets();
     void VerifyGhosts();
+
     void SaveSession();
+    void SaveInfo();
+
+    void SetStatus(const std::string &msg);
+    void SetStatus(const std::string &msg,
+                   const std::string &details);
 
     void BuildGraph(const conduit::Node &actions);
     void EnsureDomainIds();