+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Notable changes to Ascent are documented in this file. This changelog started on
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aspires to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added `save_info` action, which saves the Ascent::info() result tree at the end of execution
+
+
+
 ## [0.7.1] - Released 2021-05-20
 
 ### Preferred dependency versions for ascent@0.7.1